Rename field-resolver to resolver
